Landing Page Fixes
- For Index 4-8, sentences without periods that don't function as the
headline/catchphrase look a little unprofessional to me, but that's
just my opinion.
- Removing extraneous code from header menus
- Multiple: The "Join for free!" button being darker when we're at the
top of the page doesn't fit the site style.
- The copy on index 2 seems to imply that I would want a registry in
order to sign up ("your wedding registry").
- Index 2: The third blog post doesn't have the correct date image next
to it.
- I don't know about index 3 because there's a bunch of lorem ipsum.
- Index 4 has lorem ipsum and filler text about Mother's Day in the
campaign section, as well as the campaign names.
- REMOVED:  It also looks like it's got the same Font Awesome problem
for Facebook and Twitter, but the links for the
campaigns/facebook/twitter don't seem to point anywhere.